Modifico estructura de generación de dump
